requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
python-dotenv
pypdf
sentence-transformers
faiss-cpu
groq